benchmarks refactoring
ErrorAccSimple, ErrorAccLarge, ErrorAccNestedLarge
